uniforum-backend/perfil-service/requirements.txt
fastapi
uvicorn
python-jose
sqlalchemy
psycopg2-binary
python-dotenv
